Move non-api functions into cli module.
