+++ compaction/__init__.py
@@ -1,6 +1,6 @@
-from .compaction import compact, load_config
+from .compaction import compact
 
-__all__ = ["compact", "load_config"]
+__all__ = ["compact"]
 
 from ._version import get_versions
