+++ compaction/cli.py
@@ -1,19 +1,60 @@
+import sys
+
 import click
+import pandas
 import yaml
 
 from ._version import get_versions
-from .compaction import load_config, run_compaction
+from .compaction import compact
 
 __version__ = get_versions()["version"]
 del get_versions
 
 
+def load_config(file=None):
+    """Load compaction config file.
+
+    Parameters
+    ----------
+    fname : file-like, optional
+        Opened config file or ``None``. If ``None``, return default
+        values.
+
+    Returns
+    -------
+    dict
+        Config parameters.
+    """
+    conf = {
+        "c": 5e-8,
+        "porosity_min": 0.0,
+        "porosity_max": 1.0,
+        "rho_grain": 2650.0,
+        "rho_void": 1000.0,
+    }
+    if file is not None:
+        conf.update(yaml.load(file))
+    return conf
+
+
+def run_compaction(input=None, output=None, **kwds):
+    input = input or sys.stdin
+    output = output or sys.stdout
+
+    init = pandas.read_csv(input, names=("dz", "porosity"), dtype=float)
+
+    porosity_new = compact(init.dz.values, init.porosity.values, **kwds)
+
+    dz_new = init.dz * (1 - init.porosity) / (1 - porosity_new)
+
+    out = pandas.DataFrame.from_dict({"dz": dz_new, "porosity": porosity_new})
+    out.to_csv(output, index=False, header=False)
+
+
 @click.command()
 @click.version_option(version=__version__)
-@click.option(
-    "-v", "--verbose", is_flag=True, help="Also emit status messages to stderr."
-)
-@click.option("--dry-run", is_flag=True, help="do not actually run the model")
+@click.option("-v", "--verbose", is_flag=True, help="Emit status messages to stderr.")
+@click.option("--dry-run", is_flag=True, help="Do not actually run the model")
 @click.option(
     "--config", default=None, type=click.File(mode="r"), help="Configuration file"
 )
